Update README.md
The name of the crime is Archery King. If this game is like PUBG, an Erangel map or a place like Pochinki, and also this is a game where you take an arrow and bow from homes and play. This is a game in which 100 people land on an island. As well as. In this game you will get a lot of medkits, bandages, as well as weapons. Also, the person who survives till the end will be the winner of this game